📝 Update docs about CLI Option autocompletion, move it to its own section
