+++ docs/tutorial/options-autocompletion.md
@@ -12,10 +12,10 @@ After installing completion (for your own Python package or for **Typer CLI**), 
 
 To check it quickly without creating a new Python package, install [Typer CLI](../../typer-cli.md){.internal-link target=_blank}.
 
-Then let's create small example script:
+Then let's create small example program:
 
 ```Python
-{!../docs_src/options/autocompletion/tutorial001.py!}
+{!../docs_src/options_autocompletion/tutorial001.py!}
 ```
 
 And let's try it with **Typer CLI** to get completion:
@@ -50,8 +50,8 @@ Right now we get completion for the *CLI option* names, but not for the values.
 
 We can provide completion for the values creating an `autocompletion` function, similar to the `callback` functions from [CLI Option Callback and Context](./callback-and-context.md){.internal-link target=_blank}:
 
-```Python hl_lines="4 5  10"
-{!../docs_src/options/autocompletion/tutorial002.py!}
+```Python hl_lines="4-5  14"
+{!../docs_src/options_autocompletion/tutorial002.py!}
 ```
 
 We return a `list` of strings from the `complete_name()` function.
@@ -82,7 +82,7 @@ Modify the `complete_name()` function to receive a parameter of type `str`, it w
 Then we can check and return only the values that start with the incomplete value from the command line:
 
 ```Python hl_lines="6 7 8 9 10 11"
-{!../docs_src/options/autocompletion/tutorial003.py!}
+{!../docs_src/options_autocompletion/tutorial003.py!}
 ```
 
 Now let's try it:
@@ -120,7 +120,7 @@ In the `complete_name()` function, instead of providing one `str` per completion
 So, in the end, we return a `list` of `tuples` of `str`:
 
 ```Python hl_lines="3 4 5 6 7  10 11 12 13 14 15 16"
-{!../docs_src/options/autocompletion/tutorial004.py!}
+{!../docs_src/options_autocompletion/tutorial004.py!}
 ```
 
 !!! tip
@@ -157,7 +157,7 @@ Instead of creating and returning a list with values (`str` or `tuple`), we can 
 That way our function will be a <a href="https://docs.python.org/3.8/glossary.html#index-19" class="external-link" target="_blank">generator</a> that **Typer** (actually Click) can iterate:
 
 ```Python hl_lines="10 11 12 13"
-{!../docs_src/options/autocompletion/tutorial005.py!}
+{!../docs_src/options_autocompletion/tutorial005.py!}
 ```
 
 That simplifies our code a bit and works the same.
@@ -185,8 +185,8 @@ So, we will allow multiple `--name` *CLI options*.
 
 For this we use a `List` of `str`:
 
-```Python hl_lines="6 7 8"
-{!../docs_src/options/autocompletion/tutorial006.py!}
+```Python hl_lines="8-11"
+{!../docs_src/options_autocompletion/tutorial006.py!}
 ```
 
 And then we can use it like:
@@ -213,7 +213,7 @@ But you can access the context by declaring a function parameter of type `typer.
 And from that context you can get the current values for each parameter.
 
 ```Python hl_lines="12 13  15"
-{!../docs_src/options/autocompletion/tutorial007.py!}
+{!../docs_src/options_autocompletion/tutorial007.py!}
 ```
 
 We are getting the `names` already provided with `--name` in the command line before this completion was triggered.
@@ -282,7 +282,7 @@ You can print to "standard error" with a **Rich** `Console(stderr=True)`.
 Using `stderr=True` tells **Rich** that the output should be shown in "standard error".
 
 ```Python hl_lines="12  15-16"
-{!../docs_src/options/autocompletion/tutorial008.py!}
+{!../docs_src/options_autocompletion/tutorial008.py!}
 ```
 
 !!! info
@@ -323,7 +323,7 @@ Sebastian  -- The type hints guy.
 Of course, you can declare everything if you need it, the context, the raw *CLI parameters*, and the incomplete `str`:
 
 ```Python hl_lines="15"
-{!../docs_src/options/autocompletion/tutorial009.py!}
+{!../docs_src/options_autocompletion/tutorial009.py!}
 ```
 
 Check it:
