+++ docs_src/options_autocompletion/tutorial001.py
@@ -1,9 +1,12 @@
 import typer
 
+app = typer.Typer()
 
+
+@app.command()
 def main(name: str = typer.Option("World", help="The name to say hi to.")):
     print(f"Hello {name}")
 
 
 if __name__ == "__main__":
-    typer.run(main)
+    app()
