+++ docs_src/options_autocompletion/tutorial002.py
@@ -5,6 +5,10 @@ def complete_name():
     return ["Camila", "Carlos", "Sebastian"]
 
 
+app = typer.Typer()
+
+
+@app.command()
 def main(
     name: str = typer.Option(
         "World", help="The name to say hi to.", autocompletion=complete_name
@@ -14,4 +18,4 @@ def main(
 
 
 if __name__ == "__main__":
-    typer.run(main)
+    app()
