+++ docs_src/options_autocompletion/tutorial003.py
@@ -11,6 +11,10 @@ def complete_name(incomplete: str):
     return completion
 
 
+app = typer.Typer()
+
+
+@app.command()
 def main(
     name: str = typer.Option(
         "World", help="The name to say hi to.", autocompletion=complete_name
@@ -20,4 +24,4 @@ def main(
 
 
 if __name__ == "__main__":
-    typer.run(main)
+    app()
