+++ docs_src/options_autocompletion/tutorial004.py
@@ -16,6 +16,10 @@ def complete_name(incomplete: str):
     return completion
 
 
+app = typer.Typer()
+
+
+@app.command()
 def main(
     name: str = typer.Option(
         "World", help="The name to say hi to.", autocompletion=complete_name
@@ -25,4 +29,4 @@ def main(
 
 
 if __name__ == "__main__":
-    typer.run(main)
+    app()
