+++ docs_src/options_autocompletion/tutorial006.py
@@ -2,11 +2,14 @@ from typing import List
 
 import typer
 
+app = typer.Typer()
 
+
+@app.command()
 def main(name: List[str] = typer.Option(["World"], help="The name to say hi to.")):
     for each_name in name:
         print(f"Hello {each_name}")
 
 
 if __name__ == "__main__":
-    typer.run(main)
+    app()
