+++ docs_src/options_autocompletion/tutorial007.py
@@ -1,7 +1,6 @@
 from typing import List
 
 import typer
-from rich.console import Console
 
 valid_completion_items = [
     ("Camila", "The reader of books."),
@@ -9,16 +8,18 @@ valid_completion_items = [
     ("Sebastian", "The type hints guy."),
 ]
 
-err_console = Console(stderr=True)
 
-
-def complete_name(args: List[str], incomplete: str):
-    err_console.print(f"{args}")
+def complete_name(ctx: typer.Context, incomplete: str):
+    names = ctx.params.get("name") or []
     for name, help_text in valid_completion_items:
-        if name.startswith(incomplete):
+        if name.startswith(incomplete) and name not in names:
             yield (name, help_text)
 
 
+app = typer.Typer()
+
+
+@app.command()
 def main(
     name: List[str] = typer.Option(
         ["World"], help="The name to say hi to.", autocompletion=complete_name
@@ -29,4 +30,4 @@ def main(
 
 
 if __name__ == "__main__":
-    typer.run(main)
+    app()
