+++ mkdocs.yml
@@ -41,7 +41,6 @@ nav:
           - CLI Option Name: tutorial/options/name.md
           - CLI Option Callback and Context: tutorial/options/callback-and-context.md
           - Version CLI Option, is_eager: tutorial/options/version.md
-          - CLI Option autocompletion: tutorial/options/autocompletion.md
       - Commands:
           - Commands Intro: tutorial/commands/index.md
           - Command CLI Arguments: tutorial/commands/arguments.md
@@ -51,6 +50,7 @@ nav:
           - Typer Callback: tutorial/commands/callback.md
           - One or Multiple Commands: tutorial/commands/one-or-multiple.md
           - Using the Context: tutorial/commands/context.md
+      - CLI Option autocompletion: tutorial/options-autocompletion.md
       - CLI Parameter Types:
           - CLI Parameter Types Intro: tutorial/parameter-types/index.md
           - Number: tutorial/parameter-types/number.md
